CyllideBackend/requirements.txt
aniso8601==7.0.0
beautifulsoup4==4.8.0
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mongoengine==0.18.2
pkg-resources==0.0.0
PyJWT==1.7.1
pymongo==3.8.0
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
soupsieve==1.9.2
Werkzeug==0.15.5
